requirements.txt
Pillow==9.4.0
rich==13.3.2
